# apps/api/openapi.json
{
  "components": {
    "schemas": {
      "Announced": {
        "properties": {
          "storagePath": {
            "title": "Storagepath",
            "type": "string"
          },
          "uploadId": {
            "title": "Uploadid",
            "type": "string"
          }
        },
        "required": [
          "uploadId",
          "storagePath"
        ],
        "title": "Announced",
        "type": "object"
      },
      "AnswerSource": {
        "properties": {
          "confirmed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confirmed At"
          },
          "path": {
            "title": "Path",
            "type": "string"
          },
          "source": {
            "title": "Source",
            "type": "string"
          }
        },
        "required": [
          "path",
          "source",
          "confirmed_at"
        ],
        "title": "AnswerSource",
        "type": "object"
      },
      "ApplicationDetail": {
        "properties": {
          "answers": {
            "additionalProperties": true,
            "title": "Answers",
            "type": "object"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "destination": {
            "title": "Destination",
            "type": "string"
          },
          "draft_answers": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Draft Answers",
            "type": "object"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "last_step": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Step"
          },
          "purpose": {
            "title": "Purpose",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "submitted_job_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Submitted Job Id"
          }
        },
        "required": [
          "id",
          "destination",
          "purpose",
          "status",
          "created_at",
          "answers",
          "draft_answers",
          "last_step",
          "submitted_job_id"
        ],
        "title": "ApplicationDetail",
        "type": "object"
      },
      "ApplicationList": {
        "properties": {
          "applications": {
            "items": {
              "$ref": "#/components/schemas/ApplicationSummary"
            },
            "title": "Applications",
            "type": "array"
          }
        },
        "required": [
          "applications"
        ],
        "title": "ApplicationList",
        "type": "object"
      },
      "ApplicationSummary": {
        "properties": {
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "destination": {
            "title": "Destination",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "purpose": {
            "title": "Purpose",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "destination",
          "purpose",
          "status",
          "created_at"
        ],
        "title": "ApplicationSummary",
        "type": "object"
      },
      "ApplicationView": {
        "properties": {
          "answerSources": {
            "items": {
              "$ref": "#/components/schemas/AnswerSource"
            },
            "title": "Answersources",
            "type": "array"
          },
          "application": {
            "$ref": "#/components/schemas/ApplicationDetail"
          },
          "job": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JobState"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "application",
          "job",
          "answerSources"
        ],
        "title": "ApplicationView",
        "type": "object"
      },
      "Completeness": {
        "properties": {
          "complete": {
            "title": "Complete",
            "type": "boolean"
          },
          "missing": {
            "items": {
              "type": "string"
            },
            "title": "Missing",
            "type": "array"
          },
          "pending": {
            "items": {
              "type": "string"
            },
            "title": "Pending",
            "type": "array"
          }
        },
        "required": [
          "missing",
          "pending",
          "complete"
        ],
        "title": "Completeness",
        "type": "object"
      },
      "Created": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          }
        },
        "required": [
          "id"
        ],
        "title": "Created",
        "type": "object"
      },
      "CreatedEnvelope": {
        "properties": {
          "application": {
            "$ref": "#/components/schemas/Created"
          }
        },
        "required": [
          "application"
        ],
        "title": "CreatedEnvelope",
        "type": "object"
      },
      "DocumentsView": {
        "properties": {
          "applicationStatus": {
            "title": "Applicationstatus",
            "type": "string"
          },
          "completeness": {
            "$ref": "#/components/schemas/Completeness"
          },
          "required": {
            "items": {
              "$ref": "#/components/schemas/RequiredDocument"
            },
            "title": "Required",
            "type": "array"
          },
          "uploads": {
            "items": {
              "$ref": "#/components/schemas/Upload"
            },
            "title": "Uploads",
            "type": "array"
          }
        },
        "required": [
          "applicationStatus",
          "required",
          "uploads",
          "completeness"
        ],
        "title": "DocumentsView",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "JobState": {
        "properties": {
          "state": {
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "state"
        ],
        "title": "JobState",
        "type": "object"
      },
...
      "Position": {
        "properties": {
          "questionId": {
            "title": "Questionid",
            "type": "string"
          },
          "sectionId": {
            "title": "Sectionid",
            "type": "string"
          }
        },
        "required": [
          "sectionId",
          "questionId"
        ],
        "title": "Position",
        "type": "object"
      },
      "RequiredDocument": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "multiPage": {
            "title": "Multipage",
            "type": "boolean"
          },
          "necessity": {
            "title": "Necessity",
            "type": "string"
          }
        },
        "required": [
          "id",
          "necessity",
          "multiPage"
        ],
        "title": "RequiredDocument",
        "type": "object"
      },
...
      "Saved": {
        "properties": {
          "next": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Position"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "next"
        ],
        "title": "Saved",
        "type": "object"
      },
...
      "Upload": {
        "properties": {
          "document": {
            "title": "Document",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "original_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Original Name"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "document",
          "page",
          "original_name",
          "status"
        ],
        "title": "Upload",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "The backend every client of /api/v1 talks to (ADR-005).",
    "title": "Visa Master API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/applications": {
      "get": {
        "operationId": "list_mine_api_v1_applications_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Mine",
        "tags": [
          "applications"
        ]
      },
      "post": {
        "description": "Create a draft for a served route; 422 `route.unsupported.title` with reasons otherwise.",
        "operationId": "create_api_v1_applications_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create",
        "tags": [
          "applications"
        ]
      }
    },
    "/api/v1/applications/{application_id}": {
      "get": {
        "operationId": "get_api_v1_applications__application_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "application_id",
            "required": true,
            "schema": {
              "title": "Application Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get",
        "tags": [
          "applications"
        ]
      }
    },
    "/api/v1/applications/{application_id}/answers": {
      "post": {
        "description": "Validate and store one answer; the body says which question comes next.",
        "operationId": "save_answer_api_v1_applications__application_id__answers_post",
        "parameters": [
          {
            "in": "path",
            "name": "application_id",
            "required": true,
            "schema": {
              "title": "Application Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Saved"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Save Answer",
        "tags": [
          "applications"
        ]
      }
    },
    "/api/v1/applications/{application_id}/documents": {
      "get": {
        "operationId": "documents_api_v1_applications__application_id__documents_get",
        "parameters": [
          {
            "in": "path",
            "name": "application_id",
            "required": true,
            "schema": {
              "title": "Application Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentsView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Documents",
        "tags": [
          "applications"
        ]
      }
    },
    "/api/v1/applications/{application_id}/draft-answers": {
      "post": {
        "operationId": "save_draft_api_v1_applications__application_id__draft_answers_post",
        "parameters": [
          {
            "in": "path",
            "name": "application_id",
            "required": true,
            "schema": {
              "title": "Application Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Save Draft",
        "tags": [
          "applications"
        ]
      }
    },
    "/api/v1/applications/{application_id}/submit": {
      "post": {
        "description": "Check the whole form and its documents, then enqueue the pack — once.",
        "operationId": "submit_api_v1_applications__application_id__submit_post",
        "parameters": [
          {
            "in": "path",
            "name": "application_id",
            "required": true,
            "schema": {
              "title": "Application Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit",
        "tags": [
          "applications"
        ]
      }
    },
    "/api/v1/applications/{application_id}/uploads": {
      "post": {
        "description": "Announce an upload; the bytes then go straight to Storage at `storagePath`.",
        "operationId": "announce_api_v1_applications__application_id__uploads_post",
        "parameters": [
          {
            "in": "path",
            "name": "application_id",
            "required": true,
            "schema": {
              "title": "Application Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Announced"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Announce",
        "tags": [
          "applications"
        ]
      }
    },
    "/api/v1/applications/{application_id}/uploads/{upload_id}": {
      "delete": {
        "operationId": "remove_api_v1_applications__application_id__uploads__upload_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "application_id",
            "required": true,
            "schema": {
              "title": "Application Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "upload_id",
            "required": true,
            "schema": {
              "title": "Upload Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Remove",
        "tags": [
          "applications"
        ]
      }
    },
    "/api/v1/applications/{application_id}/uploads/{upload_id}/confirm": {
      "post": {
        "operationId": "confirm_api_v1_applications__application_id__uploads__upload_id__confirm_post",
        "parameters": [
          {
            "in": "path",
            "name": "application_id",
            "required": true,
            "schema": {
              "title": "Application Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "upload_id",
            "required": true,
            "schema": {
              "title": "Upload Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Confirm",
        "tags": [
          "applications"
        ]
      }
    },
...
  }
}